improve notes
